requirements.txt
textual>=0.38.1
mysql-connector-python>=8.0.33
redis>=4.5.4
toml>=0.10.2
cryptography>=41.0.0
pytest>=7.3.1
black>=23.3.0  # 代码格式化
isort>=5.12.0  # import排序
pylint>=2.17.0 # 代码检查